Removed the reference to constants and methods.
After a question on StackOverflow (http://stackoverflow.com/questions/13110446/accessing-constants-inside-php-object) I have removed the reference to constants and methods from the paragraph. IMO, it does not belong to a chapter about properties and the phrasing was somewhat to comprehend as well. The edit should make it easier to understand and cause less confusion.