Fix grossly misleading information
This patch has been provided by an anonymous user and solves an issue related
to the presentation of <literals> which in this case are (nearly) undistinguishable
from normal text.

Actually, we should strive for a more general solution here, since there are similar
issues in the return values section, and possibly in translations.